requirements.txt
asgiref==3.8.1
bleach==6.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.6
django-filter==24.2
django-grappelli==4.0.1
django-haystack==3.3.0
django-markdownify==0.9.5
django-seed==0.3.1
django_components==0.84
djangorestframework==3.15.2
Faker==26.0.0
idna==3.7
Markdown==3.6
packaging==24.1
pysolr==3.9.0
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
tinycss2==1.2.1
toposort==1.10
typing_extensions==4.12.2
urllib3==2.2.2
webencodings==0.5.1
